+++ tests/unit/ExceptionsBasedConditionTest.php
@@ -68,7 +68,7 @@ class ExceptionsBasedConditionTest extends TestCase
             );
             self::assertSame(
                 $expectedFailedAttempts + 1,
-                getCurrentAttempts($backoff),
+                $helper->getAttemptsMade(),
                 'total # of attempts made should be one time more than failed attempts'
             );
         }
@@ -213,7 +213,7 @@ class ExceptionsBasedConditionTest extends TestCase
             self::assertSame('an exception thrown out from class \\' . Helper::class, $t->getMessage());
             self::assertSame(
                 $maxAttempts,
-                getCurrentAttempts($backoff),
+                $helper->getAttemptsMade(),
                 'maximum number of allowed attempts have been made but all failed with exceptions thrown out'
             );
         }